Add files via upload
first of all the loading screen is fake. if you want to quickly skip the loading screen, you can just click the screen. also, there is two easter egg in options. there are two buttons in the bottom two corners of the options screen. one the left corner bottom turns the paddle to rects and the right bottom makes it so that it doesn't clear every frame. Note that the buttons are not at the exact corners but are touching. the centers of the button are at coordinate(50,550 & 850,550). look at the code if you can't find the buttons.